+++ web/src/main/java/by/pvt/kish/aircompany/command/plane/SetPlaneStatusCommand.java
@@ -21,7 +21,7 @@ public class SetPlaneStatusCommand implements by.pvt.kish.aircompany.command.Act
     public String execute(HttpServletRequest request, HttpServletResponse response) {
         String className = SetPlaneStatusCommand.class.getName();
         try {
-            int id = RequestHandler.getId(request, "pid");
+            Long id = RequestHandler.getId(request, "pid");
             PlaneService.getInstance().setStatus(id, RequestHandler.getString(request, "status"));
             request.setAttribute(Attribute.MESSAGE_ATTRIBUTE, Message.SUCCESS_SET_STATUS_PLANE);
             return Page.MAIN;
